# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InternHub API - AI-Powered Internship Management Portal
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (filtered by role context)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, interview, accepted, rejected, all]
        - name: internshipId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, faculty, company, admin, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string

        password:
          type: string
        role:
          type: string
          enum: [student, faculty, company, admin]
        university:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, faculty, company, admin]
        university:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        profileStrength:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        phone:
          type: string
        university:
          type: string
        companyName:
          type: string
        skills:
          type: array
          items:
            type: string
    Internship:
      type: object
      required: [id, role, company, location, type, duration, stipend, deadline, openings, description, tags, status, createdAt, daysLeft]
      properties:
        id:
          type: integer
        role:
          type: string
        company:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [Remote, Hybrid, On-site]
        duration:
          type: string
        stipend:
          type: string
        deadline:
          type: string
          format: date
        openings:
          type: integer
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [open, closed]
        daysLeft:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Application:
      type: object
      required: [id, internshipId, studentId, status, createdAt]
      properties:
        id:
          type: integer
        internshipId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        internship:
          $ref: "#/components/schemas/Internship"
        status:
          type: string
          enum: [pending, interview, accepted, rejected]
        coverLetter:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, interview, accepted, rejected]
...