# lib/api-zod/src/generated/types/internship.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * InternHub API - AI-Powered Internship Management Portal
 * OpenAPI spec version: 0.1.0
 */
import type { InternshipStatus } from './internshipStatus';
import type { InternshipType } from './internshipType';

export interface Internship {
  id: number;
  role: string;
  company: string;
  location: string;
  type: InternshipType;
  duration: string;
  stipend: string;
  deadline: Date;
  openings: number;
  description: string;
  tags: string[];
  status: InternshipStatus;
  daysLeft: number;
  createdAt: Date;
}